# swagger-pagamento-v1.yaml
openapi: 3.0.3
info:
  title:  vasco-bank-pagamento- OpenAPI 3.0
  description: ''
  version: 1.0.11

servers:
  - url: https:/vasco-bank-pagamento/api/v1
tags:
  - name: pagamento
    description: todos os acessos para formas de pagamentos e transferencias
    
paths:
  /boleto/{codigoBarras}:
    get:
      tags:
        - pagamento
      summary: Busca por um boleto pelo codigo de barras
      operationId: getContaCorrentById
      parameters:
        - name: codigoBarras
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: retorna um boleto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Boleto'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Boleto'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
  /boleto/{idConta}:
    post:
      tags:
        - pagamento
      summary: Pagamento de um boleto
      operationId: placeOrder
      parameters:
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: retorna o boleto pago
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Boleto'
        '405':
          description: Invalid input
          
  /transferencia/pix:
    post:
      tags:
        - pagamento
      summary: Realiza uma transferencia baseada em uma chave pix
      operationId: saldo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transferenciaDTO'
          application/xml:
            schema:
              $ref: '#/components/schemas/transferenciaDTO'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/transferenciaDTO'
        required: true
      responses:
        '200':
          description: Retorna um extrato da transferencia 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtratoConta'
        '405':
          description: Invalid input
    
components:
  schemas:
    Pagamento:
      type: object
      properties:
        tipoPagamento:
          type: string
          enum:
            - PIX
            - BOLETO
            - PAGCARTAOCREDITO
            - INTRABANCO
            - 
    
    ExtratoBoleto:
      type: object
      properties:
        id:
          type: number
          example: 24.69
        idConta:
          type: number
          example: 24.69
        numero:
          type: number
          example: 24.69
        valorCobrado:
          type: number
          example: 24.69
        valorTotal:
          type: number
          example: 24.69
        juros:
          type: number
          example: 24.69
        pago:
          type: boolean
          example: false
        dataVencimento:
          type: string
        beneficiario:
          type: string
        cpfcnpj:
          type: string
        instrucoes:
          type: string
        codigoBarras:
          type: string
        datapagamento:
          type: string
      xml:
        name: ExtratoBoleto
        
    Boleto:
      type: object
      properties:
        id:
          type: number
          example: 24.69
        numero:
          type: number
          example: 24.69
        valorCobrado:
          type: number
          example: 24.69
        valorTotal:
          type: number
          example: 24.69
        juros:
          type: number
          example: 24.69
        pago:
          type: boolean
          example: false
        dataVencimento:
          type: string
        beneficiario:
          type: string
        cpfcnpj:
          type: string
        instrucoes:
          type: string
        codigoBarras:
          type: string
        datapagamento:
          type: string
      xml:
        name: Boleto

    transferenciaDTO:
      type: object
      properties:
        chavePix: 
          type: string
        valor:
          type: number
          format: int64
          example: 99.99
      xml:
        name: transferenciaDTO
        
    ExtratoConta:
      type: object
      properties:
        idConta:
          type: number
          format: int64
          example: 1
        dataHoraPagamento: 
          type: string
        idTransicao: 
          type: string
        recebeuNome: 
          type: string
        recebeuCpf: 
          type: string
        recebeuBanco: 
          type: string
      xml:
        name: ExtratoConta
   

   
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://vasco-bank-conta/api/v1/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header